Added in a json file configuration
- Added in the ability to cache items to a json file and retreive them
  at run time
