+++ .gitignore
@@ -9,6 +9,7 @@
 *.vscode
 *.env
 *.config
+*.json
 
 # Build results
