+++ canvas-downloader.csproj
@@ -8,8 +8,16 @@
 
   <ItemGroup>
     <PackageReference Include="CommandLineParser" Version="2.8.0" />
+    <PackageReference Include="Newtonsoft.json" Version="12.0.3" />
     <PackageReference Include="System.Configuration" Version="2.0.5" />
     <PackageReference Include="System.Configuration.ConfigurationManager" Version="4.7.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+  <None Update="App.config">
+      <CopyToOutputDirectory>Never</CopyToOutputDirectory>
+      <ExcludeFromSingleFile>true</ExcludeFromSingleFile>
+  </None>
   </ItemGroup>
 
 </Project>
